Fixes as per comments 3.
