+++ src/plugins/intel_cpu/tests/functional/shared_tests_instances/skip_tests_config.cpp
@@ -198,7 +198,7 @@ std::vector<std::string> disabledTestPatterns() {
         // Issue: 123815 (Tests are sensintive to available thread count on testing machines)
         R"(.*smoke_Snippets_MHA_.?D_SplitDimensionM.*)",
         // Issue: 122356
-        R"(.*NmsRotatedLayerTestCPU.*(SortDesc=True|Clockwise=False).*)",
+        R"(.*NmsRotatedLayerTest.*(SortDesc=True|Clockwise=False).*)",
     };
 
 #if defined(OPENVINO_ARCH_X86)
