+++ src/plugins/intel_gpu/tests/functional/shared_tests_instances/single_layer_tests/nms_rotated.cpp
@@ -2,39 +2,49 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include <vector>
-
 #include "single_layer_tests/nms_rotated.hpp"
-#include "common_test_utils/test_constants.hpp"
 
 using namespace LayerTestsDefinitions;
-using namespace InferenceEngine;
-using namespace ngraph;
+using namespace ov::test;
 
-const std::vector<InputShapeParams> inShapeParams = {
-    InputShapeParams{2, 50, 50},
-    InputShapeParams {9, 10, 10}
-};
-
-const std::vector<int32_t> maxOutBoxPerClass = {5, 20};
+const std::vector<int64_t> maxOutBoxPerClass = {5, 20};
 const std::vector<float> threshold = {0.3f, 0.7f};
 const std::vector<bool> sortResDesc = {true, false};
-const std::vector<element::Type> outType = {element::i32, element::i64};
+const std::vector<ElementType> outType = {ElementType::i32, ElementType::i64};
 const std::vector<bool> clockwise = {true, false};
 
-const std::vector<Precision> inputPrecisions = {Precision::FP32, Precision::FP16};
+const std::vector<ElementType> inputPrecisions = {ElementType::f32, ElementType::f16};
 
-INSTANTIATE_TEST_SUITE_P(smoke_NmsRotatedLayerTest,
-                         NmsRotatedLayerTest,
-                         ::testing::Combine(::testing::ValuesIn(inShapeParams),
-                                            ::testing::Combine(::testing::ValuesIn(inputPrecisions),
-                                                               ::testing::Values(Precision::I32),
-                                                               ::testing::Values(Precision::FP32)),
-                                            ::testing::ValuesIn(maxOutBoxPerClass),
-                                            ::testing::ValuesIn(threshold),
-                                            ::testing::ValuesIn(threshold),
-                                            ::testing::ValuesIn(sortResDesc),
-                                            ::testing::ValuesIn(outType),
-                                            ::testing::ValuesIn(clockwise),
-                                            ::testing::Values(ov::test::utils::DEVICE_GPU)),
-                         NmsRotatedLayerTest::getTestCaseName);
+static const std::vector<std::vector<InputShape>> input_shapes = {
+    {
+        { {}, {{2, 50, 5}} },
+        { {}, {{2, 50, 50}} }
+    },
+    {
+        { {}, {{9, 10, 5}} },
+        { {}, {{9, 10, 10}} }
+    }
+};
+
+const ov::AnyMap empty_plugin_config{};
+
+INSTANTIATE_TEST_SUITE_P(smoke_NmsRotatedLayerTest, NmsRotatedLayerTest,
+        ::testing::Combine(
+                ::testing::ValuesIn(input_shapes),
+                ::testing::ValuesIn(inputPrecisions),
+                ::testing::Values(ElementType::i32),
+                ::testing::Values(ElementType::f32),
+                ::testing::ValuesIn(outType),
+                ::testing::ValuesIn(maxOutBoxPerClass),
+                ::testing::ValuesIn(threshold),
+                ::testing::ValuesIn(threshold),
+                ::testing::ValuesIn(sortResDesc),
+                ::testing::ValuesIn(clockwise),
+                ::testing::Values(false),
+                ::testing::Values(false),
+                ::testing::Values(true),
+                ::testing::Values(true),
+                ::testing::Values(true),
+                ::testing::Values(empty_plugin_config),
+                ::testing::Values(utils::DEVICE_GPU)),
+        NmsRotatedLayerTest::getTestCaseName);
