+++ src/tests/functional/plugin/shared/include/single_layer_tests/nms_rotated.hpp
@@ -9,7 +9,7 @@
 namespace LayerTestsDefinitions {
 
 TEST_P(NmsRotatedLayerTest, CompareWithRefs) {
-    Run();
+    run();
 };
 
 }  // namespace LayerTestsDefinitions
