+++ src/tests/functional/shared_test_classes/include/shared_test_classes/single_layer/nms_rotated.hpp
@@ -4,43 +4,44 @@
 
 #pragma once
 
-#include <string>
-#include <tuple>
-
-#include "shared_test_classes/base/layer_test_utils.hpp"
-
+#include "shared_test_classes/base/ov_subgraph.hpp"
 
 namespace LayerTestsDefinitions {
 
-using InputShapeParams = std::tuple<size_t,   // Number of batches
-                                    size_t,   // Number of boxes
-                                    size_t>;  // Number of classes
+typedef std::tuple<
+    std::vector<ov::test::InputShape>,  // Input shapes
+    ov::test::ElementType,              // Boxes and scores input precisions
+    ov::test::ElementType,              // Max output boxes input precisions
+    ov::test::ElementType,              // Thresholds precisions
+    ov::test::ElementType,              // Output type
+    int64_t,                            // Max output boxes per class
+    float,                              // IOU threshold
+    float,                              // Score threshold
+    bool,                               // Sort result descending
+    bool,                               // Clockwise
+    bool,                               // Is 1st input constant
+    bool,                               // Is 2nd input constant
+    bool,                               // Is 3rd input constant
+    bool,                               // Is 4th input constant
+    bool,                               // Is 5th input constant
+    ov::AnyMap,                         // Additional configuration
+    std::string                         // Device name
+> NmsRotatedParams;
 
-using InputPrecisions =
-    std::tuple<InferenceEngine::Precision,   // boxes and scores precisions
-               InferenceEngine::Precision,   // max_output_boxes_per_class precision
-               InferenceEngine::Precision>;  // iou_threshold, score_threshold, soft_nms_sigma precisions
-
-using NmsRotatedParams = std::tuple<InputShapeParams,  // Params using to create 1st and 2nd inputs
-                                    InputPrecisions,   // Input precisions
-                                    int32_t,           // Max output boxes per class
-                                    float,             // IOU threshold
-                                    float,             // Score threshold
-                                    bool,              // Sort result descending
-                                    ov::element::Type, // Output type
-                                    bool,              // Clockwise
-                                    std::string>;      // Device name
-
-class NmsRotatedLayerTest : public testing::WithParamInterface<NmsRotatedParams>, virtual public LayerTestsUtils::LayerTestsCommon {
+class NmsRotatedLayerTest : public testing::WithParamInterface<NmsRotatedParams>,
+                            public ov::test::SubgraphBaseTest {
 public:
     static std::string getTestCaseName(const testing::TestParamInfo<NmsRotatedParams>& obj);
-    void GenerateInputs() override;
-    void Compare(const std::vector<std::pair<ngraph::element::Type, std::vector<std::uint8_t>>>& expectedOutputs,
-                 const std::vector<InferenceEngine::Blob::Ptr>& actualOutputs) override;
 
 protected:
     void SetUp() override;
-    InputShapeParams inShapeParams;
+
+    void generate_inputs(const std::vector<ov::Shape>& target_shapes) override;
+
+private:
+    int64_t m_max_out_boxes_per_class;
+    float m_iou_threshold;
+    float m_score_threshold;
 };
 
-}  // namespace LayerTestsDefinitions
+} // namespace LayerTestsDefinitions
